requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
feedparser==6.0.10
Pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==2.7.2
